safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/intversion-close-in-oncomplete-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
Call db.close() in the complete handler for a version change transaction, before the success event associated with the open call fires

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-close-in-oncomplete.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
